Cambiar int por double en un get

+++ src/modelo/Producto.java
@@ -22,7 +22,7 @@ public class Producto {
         this.nombre = nombre;
     }
     
-    public double getCantidad(){
+    public int getCantidad(){
         return cantidad;
     }
